separate out test files: annotation-based vs manual
